Game Refactor
Cleaned most unnecessary things that will not make it to the final official release. Also added countdown timer. This is a "security" commit because the game will be refactored so the scoreboard isn't the main dependancy of the whole game (if there is no scoreboard the game will still be playable).